fix: コントローラのエラー管理を分離
